+++ ledger-management/src/main/java/com/midnight/midnightledger/service/TransactionService.java
@@ -8,7 +8,7 @@ import com.midnight.midnightledger.model.enums.IncomeCategory;
 import com.midnight.midnightledger.model.enums.TransactionType;
 import com.midnight.midnightledger.repository.BudgetRepository;
 import com.midnight.midnightledger.repository.TransactionRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,54 +16,44 @@ import java.math.BigDecimal;
 import java.util.*;
 
 @Service
+@RequiredArgsConstructor
 public class TransactionService {
 
     private final TransactionRepository transactionRepository;
-
     private final BudgetRepository budgetRepository;
 
-    @Autowired
-    public TransactionService(TransactionRepository transactionRepository, BudgetRepository budgetRepository) {
-        this.transactionRepository = transactionRepository;
-        this.budgetRepository = budgetRepository;
-    }
-
     @Transactional
-    public void saveTransaction(TransactionRequest transactionRequest) {
+    public void saveTransaction(TransactionRequest transactionRequest, Long accountId) {
         Transaction transaction = Transaction.builder()
                 .amount(transactionRequest.getAmount())
                 .category(transactionRequest.getCategory())
                 .transactionDate(transactionRequest.getTransactionDate())
                 .description(transactionRequest.getDescription())
                 .transactionType(transactionRequest.getTransactionType())
-                .accountId(1001L)
+                .accountId(accountId) // Associate transaction with the user
                 .build();
 
         transactionRepository.save(transaction);
 
-        Optional<Budget> budgetOptional = budgetRepository.findByCategory(transactionRequest.getCategory());
+        Optional<Budget> budgetOptional = budgetRepository.findByCategoryAndAccountId(transactionRequest.getCategory(), accountId);
 
-        if(budgetOptional.isPresent()) {
+        if (budgetOptional.isPresent()) {
             Budget budget = budgetOptional.get();
-
             BigDecimal amountUsed = budget.getAmountUsed() != null ? budget.getAmountUsed() : BigDecimal.ZERO;
-
             budget.setAmountUsed(amountUsed.add(transactionRequest.getAmount()));
-
             budgetRepository.save(budget);
         }
-
     }
 
-    public List<Transaction> getAllTransactionByTransactionType(TransactionType transactionType) {
-        return transactionRepository.findByTransactionType(transactionType).orElseThrow();
+    public List<Transaction> getAllTransactionByTransactionType(TransactionType transactionType, Long accountId) {
+        return transactionRepository.findByTransactionTypeAndAccountId(transactionType, accountId).orElseThrow();
     }
 
-    public List<Transaction> getAllTransaction() {
-        return transactionRepository.findAll();
+    public List<Transaction> getAllTransaction(Long accountId) {
+        return transactionRepository.findAllByAccountId(accountId);
     }
 
-    public Map<String, BigDecimal[]> getTransactionsByYear(int year) {
+    public Map<String, BigDecimal[]> getTransactionsByYear(int year, Long accountId) {
         BigDecimal[] monthlyIncome = new BigDecimal[12];
         BigDecimal[] monthlyExpenses = new BigDecimal[12];
         BigDecimal[] monthlySavings = new BigDecimal[12];
@@ -72,7 +62,7 @@ public class TransactionService {
         Arrays.fill(monthlyExpenses, BigDecimal.ZERO);
         Arrays.fill(monthlySavings, BigDecimal.ZERO);
 
-        List<Transaction> transactions = transactionRepository.findByYear(year);
+        List<Transaction> transactions = transactionRepository.findByYearAndAccountId(year, accountId);
 
         transactions.forEach(transaction -> {
             int monthIndex = transaction.getTransactionDate().getMonthValue() - 1; // 0-based index
@@ -93,9 +83,9 @@ public class TransactionService {
         );
     }
 
-    public Map<String, BigDecimal> getTransactionsForYearAndMonth(String transactionType, int year, int month) {
+    public Map<String, BigDecimal> getTransactionsForYearAndMonth(String transactionType, int year, int month, Long accountId) {
         TransactionType type = getTransactionType(transactionType);
-        List<Transaction> transactions = transactionRepository.findByTransactionTypeYearAndMonth(type, year, month);
+        List<Transaction> transactions = transactionRepository.findByTransactionTypeYearAndMonthAndAccountId(type, year, month, accountId);
 
         if (type == TransactionType.EXPENSES) {
             return updateMonthlyDataValue(ExpenseCategory.getExpenseMap(), transactions);
@@ -106,8 +96,8 @@ public class TransactionService {
         return Map.of(); // return empty map instead of null
     }
 
-    public Map<String, BigDecimal> getTotalPerCategories(String transactionType) {
-        List<Transaction> transactions = transactionRepository.findByTransactionType(getTransactionType(transactionType)).orElseThrow();
+    public Map<String, BigDecimal> getTotalPerCategories(String transactionType, Long accountId) {
+        List<Transaction> transactions = transactionRepository.findByTransactionTypeAndAccountId(getTransactionType(transactionType), accountId).orElseThrow();
 
         Map<String, BigDecimal> categoryTotals = ExpenseCategory.getExpenseMap();
 
@@ -125,8 +115,23 @@ public class TransactionService {
         return monthlyMap;
     }
 
-    public void deleteTransaction(Long id) {
-        transactionRepository.deleteById(id);
+
+    @Transactional
+    public void deleteTransaction(Long id, Long accountId) {
+
+        Optional<Transaction> transaction = transactionRepository.findById(id);
+
+        if(transaction.isPresent()) {
+            transactionRepository.deleteById(id);
+            Optional<Budget> budgetOptional = budgetRepository.findByCategoryAndAccountId(transaction.get().getCategory(), accountId);
+
+            if (budgetOptional.isPresent()) {
+                Budget budget = budgetOptional.get();
+                BigDecimal amountUsed = budget.getAmountUsed() != null ? budget.getAmountUsed() : BigDecimal.ZERO;
+                budget.setAmountUsed(amountUsed.subtract(transaction.get().getAmount()));
+                budgetRepository.save(budget);
+            }
+        }
     }
 
     private TransactionType getTransactionType(String transactionType) {
